+++ sanity/PortableText.tsx
@@ -1,4 +1,4 @@
-import UniversalLink from "library/link/UniversalLink"
+import UniversalLink from "library/link"
 import { styled } from "library/styled"
 import {
 	PortableText,
